@@ -4,13 +4,12 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.19.6 <3.0.0'
+  sdk: '>=3.0.2 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
   cupertino_icons: ^1.0.2
-  http: ^0.13.6
   provider: ^6.0.5
   dartz: ^0.10.1
   equatable: ^2.0.5
@@ -20,6 +19,9 @@ dependencies:
   get_it: ^7.6.0
   sqflite: ^2.2.8+4
   path_provider: ^2.0.15
+  readmore: ^2.2.0
+  intl: ^0.18.1
+  http: ^0.13.6
 
 dev_dependencies:
   flutter_test:
@@ -27,6 +29,9 @@ dev_dependencies:
   flutter_lints: ^2.0.0
   mockito: ^5.4.0
   build_runner: ^2.3.3
+  integration_test:
+    sdk: flutter
+  sqflite_common_ffi: ^2.2.5
 
 flutter:
   uses-material-design: true
