@@ -3,9 +3,9 @@ import 'package:app_ditonton/common/state_enum.dart';
 import 'package:app_ditonton/domain/entities/movie.dart';
 import 'package:app_ditonton/domain/usecases/get_movie_detail.dart';
 import 'package:app_ditonton/domain/usecases/get_movie_recommendations.dart';
-import 'package:app_ditonton/domain/usecases/get_watchlist_status.dart';
-import 'package:app_ditonton/domain/usecases/remove_watchlist.dart';
-import 'package:app_ditonton/domain/usecases/save_watchlist.dart';
+import 'package:app_ditonton/features/watchlist/domain/usecases/get_watchlist_status.dart';
+import 'package:app_ditonton/features/watchlist/domain/usecases/remove_watchlist_movie.dart';
+import 'package:app_ditonton/features/watchlist/domain/usecases/save_watchlist_movie.dart';
 import 'package:app_ditonton/presentation/provider/movie_detail_notifier.dart';
 import 'package:dartz/dartz.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -19,16 +19,16 @@ import 'movie_detail_notifier_test.mocks.dart';
   GetMovieDetail,
   GetMovieRecommendations,
   GetWatchListStatus,
-  SaveWatchlist,
-  RemoveWatchlist,
+  SaveWatchlistMovie,
+  RemoveWatchlistMovie,
 ])
 void main() {
   late MovieDetailNotifier provider;
   late MockGetMovieDetail mockGetMovieDetail;
   late MockGetMovieRecommendations mockGetMovieRecommendations;
   late MockGetWatchListStatus mockGetWatchlistStatus;
-  late MockSaveWatchlist mockSaveWatchlist;
-  late MockRemoveWatchlist mockRemoveWatchlist;
+  late MockSaveWatchlistMovie mockSaveWatchlistMovie;
+  late MockRemoveWatchlistMovie mockRemoveWatchlistMovie;
   late int listenerCallCount;
 
   setUp(() {
@@ -36,14 +36,14 @@ void main() {
     mockGetMovieDetail = MockGetMovieDetail();
     mockGetMovieRecommendations = MockGetMovieRecommendations();
     mockGetWatchlistStatus = MockGetWatchListStatus();
-    mockSaveWatchlist = MockSaveWatchlist();
-    mockRemoveWatchlist = MockRemoveWatchlist();
+    mockSaveWatchlistMovie = MockSaveWatchlistMovie();
+    mockRemoveWatchlistMovie = MockRemoveWatchlistMovie();
     provider = MovieDetailNotifier(
       getMovieDetail: mockGetMovieDetail,
       getMovieRecommendations: mockGetMovieRecommendations,
       getWatchListStatus: mockGetWatchlistStatus,
-      saveWatchlist: mockSaveWatchlist,
-      removeWatchlist: mockRemoveWatchlist,
+      saveWatchlistMovie: mockSaveWatchlistMovie,
+      removeWatchlistMovie: mockRemoveWatchlistMovie,
     )..addListener(() {
         listenerCallCount += 1;
       });
@@ -167,31 +167,44 @@ void main() {
 
     test('should execute save watchlist when function called', () async {
       // arrange
-      when(mockSaveWatchlist.execute(testMovieDetail))
+      when(mockSaveWatchlistMovie.execute(testMovieDetail))
           .thenAnswer((_) async => const Right('Success'));
       when(mockGetWatchlistStatus.execute(testMovieDetail.id))
           .thenAnswer((_) async => true);
       // act
       await provider.addWatchlist(testMovieDetail);
       // assert
-      verify(mockSaveWatchlist.execute(testMovieDetail));
+      verify(mockSaveWatchlistMovie.execute(testMovieDetail));
     });
 
     test('should execute remove watchlist when function called', () async {
       // arrange
-      when(mockRemoveWatchlist.execute(testMovieDetail))
+      when(mockRemoveWatchlistMovie.execute(testMovieDetail))
           .thenAnswer((_) async => const Right('Removed'));
       when(mockGetWatchlistStatus.execute(testMovieDetail.id))
           .thenAnswer((_) async => false);
       // act
       await provider.removeFromWatchlist(testMovieDetail);
       // assert
-      verify(mockRemoveWatchlist.execute(testMovieDetail));
+      verify(mockRemoveWatchlistMovie.execute(testMovieDetail));
+    });
+
+    test('should execute remove watchlist when function called with error',
+        () async {
+      // arrange
+      when(mockRemoveWatchlistMovie.execute(testMovieDetail))
+          .thenAnswer((_) async => const Left(ServerFailure('Error')));
+      when(mockGetWatchlistStatus.execute(testMovieDetail.id))
+          .thenAnswer((_) async => false);
+      // act
+      await provider.removeFromWatchlist(testMovieDetail);
+      // assert
+      verify(mockRemoveWatchlistMovie.execute(testMovieDetail));
     });
 
     test('should update watchlist status when add watchlist success', () async {
       // arrange
-      when(mockSaveWatchlist.execute(testMovieDetail))
+      when(mockSaveWatchlistMovie.execute(testMovieDetail))
           .thenAnswer((_) async => const Right('Added to Watchlist'));
       when(mockGetWatchlistStatus.execute(testMovieDetail.id))
           .thenAnswer((_) async => true);
@@ -206,7 +219,7 @@ void main() {
 
     test('should update watchlist message when add watchlist failed', () async {
       // arrange
-      when(mockSaveWatchlist.execute(testMovieDetail))
+      when(mockSaveWatchlistMovie.execute(testMovieDetail))
           .thenAnswer((_) async => const Left(DatabaseFailure('Failed')));
       when(mockGetWatchlistStatus.execute(testMovieDetail.id))
           .thenAnswer((_) async => false);
