@@ -1,7 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 
+const String baseURL = 'https://api.themoviedb.org/3';
+const String baseAPIKey = 'd733e78e94bdce11714e69b0946794a1';
 const String baseImageURL = 'https://image.tmdb.org/t/p/w500';
+const String imageLoadingAsset = 'assets/no-image.gif';
+const String imageLoadingURL =
+    'https://raw.githubusercontent.com/Theoarnan/app_ditonton/feature/submission_1_add_tv_series/assets/no-image.gif';
 
 // colors
 const Color kRichBlack = Color(0xFF000814);
@@ -21,23 +26,10 @@ final TextStyle kHeading6 = GoogleFonts.poppins(
   fontWeight: FontWeight.w500,
   letterSpacing: 0.15,
 );
-final TextStyle kSubtitle = GoogleFonts.poppins(
-  fontSize: 15,
-  fontWeight: FontWeight.w400,
-  letterSpacing: 0.15,
-);
 final TextStyle kBodyText = GoogleFonts.poppins(
   fontSize: 13,
   fontWeight: FontWeight.w400,
   letterSpacing: 0.25,
-);
-
-// text theme
-final kTextTheme = TextTheme(
-  titleMedium: kHeading5,
-  titleSmall: kHeading6,
-  labelMedium: kSubtitle,
-  bodyMedium: kBodyText,
 );
 
 const kColorScheme = ColorScheme(
