@@ -10,8 +10,8 @@ abstract class MovieRepository {
   Future<Either<Failure, MovieDetail>> getMovieDetail(int id);
   Future<Either<Failure, List<Movie>>> getMovieRecommendations(int id);
   Future<Either<Failure, List<Movie>>> searchMovies(String query);
-  Future<Either<Failure, String>> saveWatchlist(MovieDetail movie);
-  Future<Either<Failure, String>> removeWatchlist(MovieDetail movie);
-  Future<bool> isAddedToWatchlist(int id);
-  Future<Either<Failure, List<Movie>>> getWatchlistMovies();
+  // Future<Either<Failure, String>> saveWatchlist(MovieDetail movie);
+  // Future<Either<Failure, String>> removeWatchlist(MovieDetail movie);
+  // Future<bool> isAddedToWatchlist(int id);
+  // Future<Either<Failure, List<Movie>>> getWatchlistMovies();
 }
