@@ -2,6 +2,8 @@
 
 import 'package:app_ditonton/common/constants.dart';
 import 'package:app_ditonton/common/state_enum.dart';
+import 'package:app_ditonton/common/utils.dart';
+import 'package:app_ditonton/common/widgets.dart';
 import 'package:app_ditonton/domain/entities/genre.dart';
 import 'package:app_ditonton/domain/entities/movie.dart';
 import 'package:app_ditonton/domain/entities/movie_detail.dart';
@@ -12,11 +14,9 @@ import 'package:flutter_rating_bar/flutter_rating_bar.dart';
 import 'package:provider/provider.dart';
 
 class MovieDetailPage extends StatefulWidget {
-  static const routeName = '/detail';
+  static const routeName = '/detail-movie';
   final int id;
-
   const MovieDetailPage({super.key, required this.id});
-
   @override
   State<MovieDetailPage> createState() => _MovieDetailPageState();
 }
@@ -26,10 +26,14 @@ class _MovieDetailPageState extends State<MovieDetailPage> {
   void initState() {
     super.initState();
     Future.microtask(() {
-      Provider.of<MovieDetailNotifier>(context, listen: false)
-          .fetchMovieDetail(widget.id);
-      Provider.of<MovieDetailNotifier>(context, listen: false)
-          .loadWatchlistStatus(widget.id);
+      Provider.of<MovieDetailNotifier>(
+        context,
+        listen: false,
+      ).fetchMovieDetail(widget.id);
+      Provider.of<MovieDetailNotifier>(
+        context,
+        listen: false,
+      ).loadWatchlistStatus(widget.id);
     });
   }
 
@@ -39,12 +43,11 @@ class _MovieDetailPageState extends State<MovieDetailPage> {
       body: Consumer<MovieDetailNotifier>(
         builder: (context, provider, child) {
           if (provider.movieState == RequestState.loading) {
-            return const Center(
-              child: CircularProgressIndicator(),
-            );
+            return const Center(child: CircularProgressIndicator());
           } else if (provider.movieState == RequestState.loaded) {
             final movie = provider.movie;
             return SafeArea(
+              key: const Key('content_detail_movie'),
               child: DetailContent(
                 movie,
                 provider.movieRecommendations,
@@ -52,7 +55,10 @@ class _MovieDetailPageState extends State<MovieDetailPage> {
               ),
             );
           } else {
-            return Text(provider.message);
+            return Center(
+              key: const Key('error_message_detail_movie'),
+              child: Text(provider.message),
+            );
           }
         },
       ),
@@ -78,7 +84,7 @@ class DetailContent extends StatelessWidget {
     return Stack(
       children: [
         CachedNetworkImage(
-          imageUrl: 'https://image.tmdb.org/t/p/w500${movie.posterPath}',
+          imageUrl: '$baseImageURL${movie.posterPath}',
           width: screenWidth,
           placeholder: (context, url) => const Center(
             child: CircularProgressIndicator(),
@@ -88,8 +94,10 @@ class DetailContent extends StatelessWidget {
         Container(
           margin: const EdgeInsets.only(top: 48 + 8),
           child: DraggableScrollableSheet(
+            minChildSize: 0.25,
             builder: (context, scrollController) {
               return Container(
+                key: const Key('movieDetailScrollView'),
                 decoration: const BoxDecoration(
                   color: kRichBlack,
                   borderRadius: BorderRadius.vertical(top: Radius.circular(16)),
@@ -113,42 +121,8 @@ class DetailContent extends StatelessWidget {
                               style: kHeading5,
                             ),
                             ElevatedButton(
-                              onPressed: () async {
-                                if (!isAddedWatchlist) {
-                                  await Provider.of<MovieDetailNotifier>(
-                                          context,
-                                          listen: false)
-                                      .addWatchlist(movie);
-                                } else {
-                                  await Provider.of<MovieDetailNotifier>(
-                                          context,
-                                          listen: false)
-                                      .removeFromWatchlist(movie);
-                                }
-
-                                final message =
-                                    Provider.of<MovieDetailNotifier>(context,
-                                            listen: false)
-                                        .watchlistMessage;
-
-                                if (message ==
-                                        MovieDetailNotifier
-                                            .watchlistAddSuccessMessage ||
-                                    message ==
-                                        MovieDetailNotifier
-                                            .watchlistRemoveSuccessMessage) {
-                                  ScaffoldMessenger.of(context).showSnackBar(
-                                      SnackBar(content: Text(message)));
-                                } else {
-                                  showDialog(
-                                      context: context,
-                                      builder: (context) {
-                                        return AlertDialog(
-                                          content: Text(message),
-                                        );
-                                      });
-                                }
-                              },
+                              key: const Key('moveButtonWatchlist'),
+                              onPressed: () => handleOnTapButton(context),
                               child: Row(
                                 mainAxisSize: MainAxisSize.min,
                                 children: [
@@ -159,12 +133,8 @@ class DetailContent extends StatelessWidget {
                                 ],
                               ),
                             ),
-                            Text(
-                              _showGenres(movie.genres),
-                            ),
-                            Text(
-                              _showDuration(movie.runtime),
-                            ),
+                            Text(_showGenres(movie.genres)),
+                            Text(_showDuration(movie.runtime)),
                             Row(
                               children: [
                                 RatingBarIndicator(
@@ -176,17 +146,12 @@ class DetailContent extends StatelessWidget {
                                   ),
                                   itemSize: 24,
                                 ),
-                                Text('${movie.voteAverage}')
+                                Text(formatVoteAverage(movie.voteAverage))
                               ],
                             ),
                             const SizedBox(height: 16),
-                            Text(
-                              'Overview',
-                              style: kHeading6,
-                            ),
-                            Text(
-                              movie.overview,
-                            ),
+                            Text('Overview', style: kHeading6),
+                            Text(movie.overview),
                             const SizedBox(height: 16),
                             Text(
                               'Recommendations',
@@ -194,20 +159,28 @@ class DetailContent extends StatelessWidget {
                             ),
                             Consumer<MovieDetailNotifier>(
                               builder: (context, data, child) {
-                                if (data.recommendationState ==
-                                    RequestState.loading) {
+                                final state = data.recommendationState;
+                                if (state == RequestState.loading) {
                                   return const Center(
                                     child: CircularProgressIndicator(),
                                   );
-                                } else if (data.recommendationState ==
-                                    RequestState.error) {
-                                  return Text(data.message);
-                                } else if (data.recommendationState ==
-                                    RequestState.loaded) {
+                                } else if (state == RequestState.error) {
+                                  return Center(
+                                    key: const Key('error_message_2'),
+                                    child: Text(data.message),
+                                  );
+                                } else if (state == RequestState.loaded) {
                                   return SizedBox(
+                                    key: const Key(
+                                      'content_movie_detail_recommendation',
+                                    ),
                                     height: 150,
                                     child: ListView.builder(
+                                      key: const Key(
+                                        'list_builder_recomendation',
+                                      ),
                                       scrollDirection: Axis.horizontal,
+                                      itemCount: recommendations.length,
                                       itemBuilder: (context, index) {
                                         final movie = recommendations[index];
                                         return Padding(
@@ -220,32 +193,16 @@ class DetailContent extends StatelessWidget {
                                                 arguments: movie.id,
                                               );
                                             },
-                                            child: ClipRRect(
-                                              borderRadius:
-                                                  const BorderRadius.all(
-                                                Radius.circular(8),
-                                              ),
-                                              child: CachedNetworkImage(
-                                                imageUrl:
-                                                    'https://image.tmdb.org/t/p/w500${movie.posterPath}',
-                                                placeholder: (context, url) =>
-                                                    const Center(
-                                                  child:
-                                                      CircularProgressIndicator(),
-                                                ),
-                                                errorWidget:
-                                                    (context, url, error) =>
-                                                        const Icon(Icons.error),
-                                              ),
-                                            ),
+                                            child: cardRecomendation(movie),
                                           ),
                                         );
                                       },
-                                      itemCount: recommendations.length,
                                     ),
                                   );
                                 } else {
-                                  return Container();
+                                  return const SizedBox(
+                                    key: Key('empty_widget'),
+                                  );
                                 }
                               },
                             ),
@@ -265,9 +222,6 @@ class DetailContent extends StatelessWidget {
                 ),
               );
             },
-            // initialChildSize: 0.5,
-            minChildSize: 0.25,
-            // maxChildSize: 1.0,
           ),
         ),
         Padding(
@@ -276,6 +230,7 @@ class DetailContent extends StatelessWidget {
             backgroundColor: kRichBlack,
             foregroundColor: Colors.white,
             child: IconButton(
+              key: const Key('buttonBackDetailMovie'),
               icon: const Icon(Icons.arrow_back),
               onPressed: () {
                 Navigator.pop(context);
@@ -287,16 +242,57 @@ class DetailContent extends StatelessWidget {
     );
   }
 
+  Widget cardRecomendation(Movie movie) {
+    return Widgets.imageCachedNetworkCard(movie.posterPath);
+  }
+
+  Future<void> handleOnTapButton(BuildContext context) async {
+    if (!isAddedWatchlist) {
+      await Provider.of<MovieDetailNotifier>(
+        context,
+        listen: false,
+      ).addWatchlist(
+        movie,
+      );
+    } else {
+      await Provider.of<MovieDetailNotifier>(
+        context,
+        listen: false,
+      ).removeFromWatchlist(
+        movie,
+      );
+    }
+
+    final message = Provider.of<MovieDetailNotifier>(
+      context,
+      listen: false,
+    ).watchlistMessage;
+
+    if (message == MovieDetailNotifier.watchlistAddSuccessMessage ||
+        message == MovieDetailNotifier.watchlistRemoveSuccessMessage) {
+      ScaffoldMessenger.of(context).showSnackBar(
+        SnackBar(content: Text(message)),
+      );
+    } else {
+      showDialog(
+        context: context,
+        builder: (context) {
+          return AlertDialog(
+            content: Text(message),
+          );
+        },
+      );
+    }
+  }
+
   String _showGenres(List<Genre> genres) {
     String result = '';
     for (var genre in genres) {
       result += '${genre.name}, ';
     }
-
     if (result.isEmpty) {
       return result;
     }
-
     return result.substring(0, result.length - 2);
   }
 
