@@ -6,9 +6,7 @@ import 'package:provider/provider.dart';
 
 class TopRatedMoviesPage extends StatefulWidget {
   static const routeName = '/top-rated-movie';
-
   const TopRatedMoviesPage({super.key});
-
   @override
   State<TopRatedMoviesPage> createState() => _TopRatedMoviesPageState();
 }
@@ -17,9 +15,12 @@ class _TopRatedMoviesPageState extends State<TopRatedMoviesPage> {
   @override
   void initState() {
     super.initState();
-    Future.microtask(() =>
-        Provider.of<TopRatedMoviesNotifier>(context, listen: false)
-            .fetchTopRatedMovies());
+    Future.microtask(
+      () => Provider.of<TopRatedMoviesNotifier>(
+        context,
+        listen: false,
+      ).fetchTopRatedMovies(),
+    );
   }
 
   @override
@@ -38,9 +39,13 @@ class _TopRatedMoviesPageState extends State<TopRatedMoviesPage> {
               );
             } else if (data.state == RequestState.loaded) {
               return ListView.builder(
+                key: const Key('listTopRatedMovies'),
                 itemBuilder: (context, index) {
                   final movie = data.movies[index];
-                  return MovieCard(movie);
+                  return MovieCard(
+                    key: Key('listTopRatedMovies$index'),
+                    movie,
+                  );
                 },
                 itemCount: data.movies.length,
               );
