@@ -3,9 +3,9 @@ import 'package:app_ditonton/domain/entities/movie.dart';
 import 'package:app_ditonton/domain/entities/movie_detail.dart';
 import 'package:app_ditonton/domain/usecases/get_movie_detail.dart';
 import 'package:app_ditonton/domain/usecases/get_movie_recommendations.dart';
-import 'package:app_ditonton/domain/usecases/get_watchlist_status.dart';
-import 'package:app_ditonton/domain/usecases/remove_watchlist.dart';
-import 'package:app_ditonton/domain/usecases/save_watchlist.dart';
+import 'package:app_ditonton/features/watchlist/domain/usecases/get_watchlist_status.dart';
+import 'package:app_ditonton/features/watchlist/domain/usecases/remove_watchlist_movie.dart';
+import 'package:app_ditonton/features/watchlist/domain/usecases/save_watchlist_movie.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
@@ -16,15 +16,15 @@ class MovieDetailNotifier extends ChangeNotifier {
   final GetMovieDetail getMovieDetail;
   final GetMovieRecommendations getMovieRecommendations;
   final GetWatchListStatus getWatchListStatus;
-  final SaveWatchlist saveWatchlist;
-  final RemoveWatchlist removeWatchlist;
+  final SaveWatchlistMovie saveWatchlistMovie;
+  final RemoveWatchlistMovie removeWatchlistMovie;
 
   MovieDetailNotifier({
     required this.getMovieDetail,
     required this.getMovieRecommendations,
     required this.getWatchListStatus,
-    required this.saveWatchlist,
-    required this.removeWatchlist,
+    required this.saveWatchlistMovie,
+    required this.removeWatchlistMovie,
   });
 
   late MovieDetail _movie;
@@ -80,8 +80,7 @@ class MovieDetailNotifier extends ChangeNotifier {
   String get watchlistMessage => _watchlistMessage;
 
   Future<void> addWatchlist(MovieDetail movie) async {
-    final result = await saveWatchlist.execute(movie);
-
+    final result = await saveWatchlistMovie.execute(movie);
     await result.fold(
       (failure) async {
         _watchlistMessage = failure.message;
@@ -90,13 +89,11 @@ class MovieDetailNotifier extends ChangeNotifier {
         _watchlistMessage = successMessage;
       },
     );
-
     await loadWatchlistStatus(movie.id);
   }
 
   Future<void> removeFromWatchlist(MovieDetail movie) async {
-    final result = await removeWatchlist.execute(movie);
-
+    final result = await removeWatchlistMovie.execute(movie);
     await result.fold(
       (failure) async {
         _watchlistMessage = failure.message;
@@ -105,7 +102,6 @@ class MovieDetailNotifier extends ChangeNotifier {
         _watchlistMessage = successMessage;
       },
     );
-
     await loadWatchlistStatus(movie.id);
   }
 
