@@ -1,6 +1,8 @@
-// Mocks generated by Mockito 5.4.0 from annotations
+// Mocks generated by Mockito 5.4.1 from annotations
 // in app_ditonton/test/presentation/pages/movie_detail_page_test.dart.
 // Do not manually edit this file.
+
+// @dart=2.19
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'dart:async' as _i11;
@@ -12,9 +14,12 @@ import 'package:app_ditonton/domain/entities/movie_detail.dart' as _i7;
 import 'package:app_ditonton/domain/usecases/get_movie_detail.dart' as _i2;
 import 'package:app_ditonton/domain/usecases/get_movie_recommendations.dart'
     as _i3;
-import 'package:app_ditonton/domain/usecases/get_watchlist_status.dart' as _i4;
-import 'package:app_ditonton/domain/usecases/remove_watchlist.dart' as _i6;
-import 'package:app_ditonton/domain/usecases/save_watchlist.dart' as _i5;
+import 'package:app_ditonton/features/watchlist/domain/usecases/get_watchlist_status.dart'
+    as _i4;
+import 'package:app_ditonton/features/watchlist/domain/usecases/remove_watchlist_movie.dart'
+    as _i6;
+import 'package:app_ditonton/features/watchlist/domain/usecases/save_watchlist_movie.dart'
+    as _i5;
 import 'package:app_ditonton/presentation/provider/movie_detail_notifier.dart'
     as _i8;
 import 'package:mockito/mockito.dart' as _i1;
@@ -63,8 +68,9 @@ class _FakeGetWatchListStatus_2 extends _i1.SmartFake
         );
 }
 
-class _FakeSaveWatchlist_3 extends _i1.SmartFake implements _i5.SaveWatchlist {
-  _FakeSaveWatchlist_3(
+class _FakeSaveWatchlistMovie_3 extends _i1.SmartFake
+    implements _i5.SaveWatchlistMovie {
+  _FakeSaveWatchlistMovie_3(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -73,9 +79,9 @@ class _FakeSaveWatchlist_3 extends _i1.SmartFake implements _i5.SaveWatchlist {
         );
 }
 
-class _FakeRemoveWatchlist_4 extends _i1.SmartFake
-    implements _i6.RemoveWatchlist {
-  _FakeRemoveWatchlist_4(
+class _FakeRemoveWatchlistMovie_4 extends _i1.SmartFake
+    implements _i6.RemoveWatchlistMovie {
+  _FakeRemoveWatchlistMovie_4(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -129,21 +135,21 @@ class MockMovieDetailNotifier extends _i1.Mock
         ),
       ) as _i4.GetWatchListStatus);
   @override
-  _i5.SaveWatchlist get saveWatchlist => (super.noSuchMethod(
-        Invocation.getter(#saveWatchlist),
-        returnValue: _FakeSaveWatchlist_3(
+  _i5.SaveWatchlistMovie get saveWatchlistMovie => (super.noSuchMethod(
+        Invocation.getter(#saveWatchlistMovie),
+        returnValue: _FakeSaveWatchlistMovie_3(
           this,
-          Invocation.getter(#saveWatchlist),
+          Invocation.getter(#saveWatchlistMovie),
         ),
-      ) as _i5.SaveWatchlist);
+      ) as _i5.SaveWatchlistMovie);
   @override
-  _i6.RemoveWatchlist get removeWatchlist => (super.noSuchMethod(
-        Invocation.getter(#removeWatchlist),
-        returnValue: _FakeRemoveWatchlist_4(
+  _i6.RemoveWatchlistMovie get removeWatchlistMovie => (super.noSuchMethod(
+        Invocation.getter(#removeWatchlistMovie),
+        returnValue: _FakeRemoveWatchlistMovie_4(
           this,
-          Invocation.getter(#removeWatchlist),
+          Invocation.getter(#removeWatchlistMovie),
         ),
-      ) as _i6.RemoveWatchlist);
+      ) as _i6.RemoveWatchlistMovie);
   @override
   _i7.MovieDetail get movie => (super.noSuchMethod(
         Invocation.getter(#movie),
