@@ -1,5 +1,4 @@
 import 'package:app_ditonton/common/state_enum.dart';
-import 'package:app_ditonton/domain/entities/movie.dart';
 import 'package:app_ditonton/presentation/pages/popular_movies_page.dart';
 import 'package:app_ditonton/presentation/provider/popular_movies_notifier.dart';
 import 'package:flutter/material.dart';
@@ -8,6 +7,7 @@ import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
 import 'package:provider/provider.dart';
 
+import '../../dummy_data/dummy_objects.dart';
 import 'popular_movies_page_test.mocks.dart';
 
 @GenerateMocks([PopularMoviesNotifier])
@@ -27,40 +27,42 @@ void main() {
     );
   }
 
-  testWidgets('Page should display center progress bar when loading',
-      (WidgetTester tester) async {
-    when(mockNotifier.state).thenReturn(RequestState.loading);
+  group('Page Popular Movie', () {
+    testWidgets('Page should display center progress bar when loading',
+        (WidgetTester tester) async {
+      when(mockNotifier.state).thenReturn(RequestState.loading);
 
-    final progressBarFinder = find.byType(CircularProgressIndicator);
-    final centerFinder = find.byType(Center);
+      final progressBarFinder = find.byType(CircularProgressIndicator);
+      final centerFinder = find.byType(Center);
 
-    await tester.pumpWidget(makeTestableWidget(const PopularMoviesPage()));
+      await tester.pumpWidget(makeTestableWidget(const PopularMoviesPage()));
 
-    expect(centerFinder, findsOneWidget);
-    expect(progressBarFinder, findsOneWidget);
-  });
+      expect(centerFinder, findsOneWidget);
+      expect(progressBarFinder, findsOneWidget);
+    });
 
-  testWidgets('Page should display ListView when data is loaded',
-      (WidgetTester tester) async {
-    when(mockNotifier.state).thenReturn(RequestState.loaded);
-    when(mockNotifier.movies).thenReturn(<Movie>[]);
+    testWidgets('Page should display ListView when data is loaded',
+        (WidgetTester tester) async {
+      when(mockNotifier.state).thenReturn(RequestState.loaded);
+      when(mockNotifier.movies).thenReturn(testMovieList);
 
-    final listViewFinder = find.byType(ListView);
+      final listViewFinder = find.byType(ListView);
 
-    await tester.pumpWidget(makeTestableWidget(const PopularMoviesPage()));
+      await tester.pumpWidget(makeTestableWidget(const PopularMoviesPage()));
 
-    expect(listViewFinder, findsOneWidget);
-  });
+      expect(listViewFinder, findsOneWidget);
+    });
 
-  testWidgets('Page should display text with message when Error',
-      (WidgetTester tester) async {
-    when(mockNotifier.state).thenReturn(RequestState.error);
-    when(mockNotifier.message).thenReturn('Error message');
+    testWidgets('Page should display text with message when Error',
+        (WidgetTester tester) async {
+      when(mockNotifier.state).thenReturn(RequestState.error);
+      when(mockNotifier.message).thenReturn('Error message');
 
-    final textFinder = find.byKey(const Key('error_message'));
+      final textFinder = find.byKey(const Key('error_message'));
 
-    await tester.pumpWidget(makeTestableWidget(const PopularMoviesPage()));
+      await tester.pumpWidget(makeTestableWidget(const PopularMoviesPage()));
 
-    expect(textFinder, findsOneWidget);
+      expect(textFinder, findsOneWidget);
+    });
   });
 }
