@@ -1,6 +1,8 @@
-// Mocks generated by Mockito 5.4.0 from annotations
+// Mocks generated by Mockito 5.4.1 from annotations
 // in app_ditonton/test/presentation/pages/popular_movies_page_test.dart.
 // Do not manually edit this file.
+
+// @dart=2.19
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'dart:async' as _i6;
