@@ -1,5 +1,4 @@
 import 'package:app_ditonton/common/state_enum.dart';
-import 'package:app_ditonton/domain/entities/movie.dart';
 import 'package:app_ditonton/presentation/pages/top_rated_movies_page.dart';
 import 'package:app_ditonton/presentation/provider/top_rated_movies_notifier.dart';
 import 'package:flutter/material.dart';
@@ -8,6 +7,7 @@ import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
 import 'package:provider/provider.dart';
 
+import '../../dummy_data/dummy_objects.dart';
 import 'top_rated_movies_page_test.mocks.dart';
 
 @GenerateMocks([TopRatedMoviesNotifier])
@@ -43,7 +43,7 @@ void main() {
   testWidgets('Page should display when data is loaded',
       (WidgetTester tester) async {
     when(mockNotifier.state).thenReturn(RequestState.loaded);
-    when(mockNotifier.movies).thenReturn(<Movie>[]);
+    when(mockNotifier.movies).thenReturn(testMovieList);
 
     final listViewFinder = find.byType(ListView);
 
