@@ -1,20 +1,27 @@
-// Mocks generated by Mockito 5.4.0 from annotations
+// Mocks generated by Mockito 5.4.1 from annotations
 // in app_ditonton/test/presentation/provider/movie_detail_notifier_test.dart.
 // Do not manually edit this file.
 
-// ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'dart:async' as _i5;
+// @dart=2.19
 
-import 'package:app_ditonton/common/failure.dart' as _i6;
-import 'package:app_ditonton/domain/entities/movie.dart' as _i9;
-import 'package:app_ditonton/domain/entities/movie_detail.dart' as _i7;
+// ignore_for_file: no_leading_underscores_for_library_prefixes
+import 'dart:async' as _i6;
+
+import 'package:app_ditonton/common/failure.dart' as _i7;
+import 'package:app_ditonton/domain/entities/movie.dart' as _i10;
+import 'package:app_ditonton/domain/entities/movie_detail.dart' as _i8;
 import 'package:app_ditonton/domain/repositories/movie_repository.dart' as _i2;
-import 'package:app_ditonton/domain/usecases/get_movie_detail.dart' as _i4;
+import 'package:app_ditonton/domain/usecases/get_movie_detail.dart' as _i5;
 import 'package:app_ditonton/domain/usecases/get_movie_recommendations.dart'
-    as _i8;
-import 'package:app_ditonton/domain/usecases/get_watchlist_status.dart' as _i10;
-import 'package:app_ditonton/domain/usecases/remove_watchlist.dart' as _i12;
-import 'package:app_ditonton/domain/usecases/save_watchlist.dart' as _i11;
+    as _i9;
+import 'package:app_ditonton/features/watchlist/domain/repositories/watchlist_repository.dart'
+    as _i4;
+import 'package:app_ditonton/features/watchlist/domain/usecases/get_watchlist_status.dart'
+    as _i11;
+import 'package:app_ditonton/features/watchlist/domain/usecases/remove_watchlist_movie.dart'
+    as _i13;
+import 'package:app_ditonton/features/watchlist/domain/usecases/save_watchlist_movie.dart'
+    as _i12;
 import 'package:dartz/dartz.dart' as _i3;
 import 'package:mockito/mockito.dart' as _i1;
 
@@ -50,10 +57,21 @@ class _FakeEither_1<L, R> extends _i1.SmartFake implements _i3.Either<L, R> {
         );
 }
 
+class _FakeWatchlistRepository_2 extends _i1.SmartFake
+    implements _i4.WatchlistRepository {
+  _FakeWatchlistRepository_2(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
+}
+
 /// A class which mocks [GetMovieDetail].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockGetMovieDetail extends _i1.Mock implements _i4.GetMovieDetail {
+class MockGetMovieDetail extends _i1.Mock implements _i5.GetMovieDetail {
   MockGetMovieDetail() {
     _i1.throwOnMissingStub(this);
   }
@@ -67,28 +85,28 @@ class MockGetMovieDetail extends _i1.Mock implements _i4.GetMovieDetail {
         ),
       ) as _i2.MovieRepository);
   @override
-  _i5.Future<_i3.Either<_i6.Failure, _i7.MovieDetail>> execute(int? id) =>
+  _i6.Future<_i3.Either<_i7.Failure, _i8.MovieDetail>> execute(int? id) =>
       (super.noSuchMethod(
         Invocation.method(
           #execute,
           [id],
         ),
-        returnValue: _i5.Future<_i3.Either<_i6.Failure, _i7.MovieDetail>>.value(
-            _FakeEither_1<_i6.Failure, _i7.MovieDetail>(
+        returnValue: _i6.Future<_i3.Either<_i7.Failure, _i8.MovieDetail>>.value(
+            _FakeEither_1<_i7.Failure, _i8.MovieDetail>(
           this,
           Invocation.method(
             #execute,
             [id],
           ),
         )),
-      ) as _i5.Future<_i3.Either<_i6.Failure, _i7.MovieDetail>>);
+      ) as _i6.Future<_i3.Either<_i7.Failure, _i8.MovieDetail>>);
 }
 
 /// A class which mocks [GetMovieRecommendations].
 ///
 /// See the documentation for Mockito's code generation for more information.
 class MockGetMovieRecommendations extends _i1.Mock
-    implements _i8.GetMovieRecommendations {
+    implements _i9.GetMovieRecommendations {
   MockGetMovieRecommendations() {
     _i1.throwOnMissingStub(this);
   }
@@ -102,114 +120,117 @@ class MockGetMovieRecommendations extends _i1.Mock
         ),
       ) as _i2.MovieRepository);
   @override
-  _i5.Future<_i3.Either<_i6.Failure, List<_i9.Movie>>> execute(dynamic id) =>
+  _i6.Future<_i3.Either<_i7.Failure, List<_i10.Movie>>> execute(dynamic id) =>
       (super.noSuchMethod(
         Invocation.method(
           #execute,
           [id],
         ),
-        returnValue: _i5.Future<_i3.Either<_i6.Failure, List<_i9.Movie>>>.value(
-            _FakeEither_1<_i6.Failure, List<_i9.Movie>>(
+        returnValue:
+            _i6.Future<_i3.Either<_i7.Failure, List<_i10.Movie>>>.value(
+                _FakeEither_1<_i7.Failure, List<_i10.Movie>>(
           this,
           Invocation.method(
             #execute,
             [id],
           ),
         )),
-      ) as _i5.Future<_i3.Either<_i6.Failure, List<_i9.Movie>>>);
+      ) as _i6.Future<_i3.Either<_i7.Failure, List<_i10.Movie>>>);
 }
 
 /// A class which mocks [GetWatchListStatus].
 ///
 /// See the documentation for Mockito's code generation for more information.
 class MockGetWatchListStatus extends _i1.Mock
-    implements _i10.GetWatchListStatus {
+    implements _i11.GetWatchListStatus {
   MockGetWatchListStatus() {
     _i1.throwOnMissingStub(this);
   }
 
   @override
-  _i2.MovieRepository get repository => (super.noSuchMethod(
+  _i4.WatchlistRepository get repository => (super.noSuchMethod(
         Invocation.getter(#repository),
-        returnValue: _FakeMovieRepository_0(
+        returnValue: _FakeWatchlistRepository_2(
           this,
           Invocation.getter(#repository),
         ),
-      ) as _i2.MovieRepository);
+      ) as _i4.WatchlistRepository);
   @override
-  _i5.Future<bool> execute(int? id) => (super.noSuchMethod(
+  _i6.Future<bool> execute(int? id) => (super.noSuchMethod(
         Invocation.method(
           #execute,
           [id],
         ),
-        returnValue: _i5.Future<bool>.value(false),
-      ) as _i5.Future<bool>);
+        returnValue: _i6.Future<bool>.value(false),
+      ) as _i6.Future<bool>);
 }
 
-/// A class which mocks [SaveWatchlist].
+/// A class which mocks [SaveWatchlistMovie].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockSaveWatchlist extends _i1.Mock implements _i11.SaveWatchlist {
-  MockSaveWatchlist() {
+class MockSaveWatchlistMovie extends _i1.Mock
+    implements _i12.SaveWatchlistMovie {
+  MockSaveWatchlistMovie() {
     _i1.throwOnMissingStub(this);
   }
 
   @override
-  _i2.MovieRepository get repository => (super.noSuchMethod(
+  _i4.WatchlistRepository get repository => (super.noSuchMethod(
         Invocation.getter(#repository),
-        returnValue: _FakeMovieRepository_0(
+        returnValue: _FakeWatchlistRepository_2(
           this,
           Invocation.getter(#repository),
         ),
-      ) as _i2.MovieRepository);
+      ) as _i4.WatchlistRepository);
   @override
-  _i5.Future<_i3.Either<_i6.Failure, String>> execute(_i7.MovieDetail? movie) =>
+  _i6.Future<_i3.Either<_i7.Failure, String>> execute(_i8.MovieDetail? movie) =>
       (super.noSuchMethod(
         Invocation.method(
           #execute,
           [movie],
         ),
-        returnValue: _i5.Future<_i3.Either<_i6.Failure, String>>.value(
-            _FakeEither_1<_i6.Failure, String>(
+        returnValue: _i6.Future<_i3.Either<_i7.Failure, String>>.value(
+            _FakeEither_1<_i7.Failure, String>(
           this,
           Invocation.method(
             #execute,
             [movie],
           ),
         )),
-      ) as _i5.Future<_i3.Either<_i6.Failure, String>>);
+      ) as _i6.Future<_i3.Either<_i7.Failure, String>>);
 }
 
-/// A class which mocks [RemoveWatchlist].
+/// A class which mocks [RemoveWatchlistMovie].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockRemoveWatchlist extends _i1.Mock implements _i12.RemoveWatchlist {
-  MockRemoveWatchlist() {
+class MockRemoveWatchlistMovie extends _i1.Mock
+    implements _i13.RemoveWatchlistMovie {
+  MockRemoveWatchlistMovie() {
     _i1.throwOnMissingStub(this);
   }
 
   @override
-  _i2.MovieRepository get repository => (super.noSuchMethod(
+  _i4.WatchlistRepository get repository => (super.noSuchMethod(
         Invocation.getter(#repository),
-        returnValue: _FakeMovieRepository_0(
+        returnValue: _FakeWatchlistRepository_2(
           this,
           Invocation.getter(#repository),
         ),
-      ) as _i2.MovieRepository);
+      ) as _i4.WatchlistRepository);
   @override
-  _i5.Future<_i3.Either<_i6.Failure, String>> execute(_i7.MovieDetail? movie) =>
+  _i6.Future<_i3.Either<_i7.Failure, String>> execute(_i8.MovieDetail? movie) =>
       (super.noSuchMethod(
         Invocation.method(
           #execute,
           [movie],
         ),
-        returnValue: _i5.Future<_i3.Either<_i6.Failure, String>>.value(
-            _FakeEither_1<_i6.Failure, String>(
+        returnValue: _i6.Future<_i3.Either<_i7.Failure, String>>.value(
+            _FakeEither_1<_i7.Failure, String>(
           this,
           Invocation.method(
             #execute,
             [movie],
           ),
         )),
-      ) as _i5.Future<_i3.Either<_i6.Failure, String>>);
+      ) as _i6.Future<_i3.Either<_i7.Failure, String>>);
 }
